UPnp: Move Format.* functions to HTTPRequest
