+++ mythtv/libs/libmythupnp/httprequest.cpp
@@ -53,10 +53,68 @@
 #include "serializers/xmlplistSerializer.h"
 
 #include "httpserver.h"
+#include "upnpresultcode.h"
 
 #ifndef O_LARGEFILE
 #define O_LARGEFILE 0
 #endif
+
+QString UPnPResultCodeDesc(UPnPResultCode eCode)
+{
+    switch( eCode )
+    {
+        case UPnPResult_Success                     : return "Success";
+        case UPnPResult_InvalidAction               : return "Invalid Action";
+        case UPnPResult_InvalidArgs                 : return "Invalid Args";
+        case UPnPResult_ActionFailed                : return "Action Failed";
+        case UPnPResult_ArgumentValueInvalid        : return "Argument Value Invalid";
+        case UPnPResult_ArgumentValueOutOfRange     : return "Argument Value Out Of Range";
+        case UPnPResult_OptionalActionNotImplemented: return "Optional Action Not Implemented";
+        case UPnPResult_OutOfMemory                 : return "Out Of Memory";
+        case UPnPResult_HumanInterventionRequired   : return "Human Intervention Required";
+        case UPnPResult_StringArgumentTooLong       : return "String Argument Too Long";
+        case UPnPResult_ActionNotAuthorized         : return "Action Not Authorized";
+        case UPnPResult_SignatureFailure            : return "Signature Failure";
+        case UPnPResult_SignatureMissing            : return "Signature Missing";
+        case UPnPResult_NotEncrypted                : return "Not Encrypted";
+        case UPnPResult_InvalidSequence             : return "Invalid Sequence";
+        case UPnPResult_InvalidControlURL           : return "Invalid Control URL";
+        case UPnPResult_NoSuchSession               : return "No Such Session";
+        case UPnPResult_MS_AccessDenied             : return "Access Denied";
+
+        case UPnPResult_CDS_NoSuchObject            : return "No Such Object";
+        case UPnPResult_CDS_InvalidCurrentTagValue  : return "Invalid CurrentTagValue";
+        case UPnPResult_CDS_InvalidNewTagValue      : return "Invalid NewTagValue";
+        case UPnPResult_CDS_RequiredTag             : return "Required Tag";
+        case UPnPResult_CDS_ReadOnlyTag             : return "Read Only Tag";
+        case UPnPResult_CDS_ParameterMismatch       : return "Parameter Mismatch";
+        case UPnPResult_CDS_InvalidSearchCriteria   : return "Invalid Search Criteria";
+        case UPnPResult_CDS_InvalidSortCriteria     : return "Invalid Sort Criteria";
+        case UPnPResult_CDS_NoSuchContainer         : return "No Such Container";
+        case UPnPResult_CDS_RestrictedObject        : return "Restricted Object";
+        case UPnPResult_CDS_BadMetadata             : return "Bad Metadata";
+        case UPnPResult_CDS_ResrtictedParentObject  : return "Resrticted Parent Object";
+        case UPnPResult_CDS_NoSuchSourceResource    : return "No Such Source Resource";
+        case UPnPResult_CDS_ResourceAccessDenied    : return "Resource Access Denied";
+        case UPnPResult_CDS_TransferBusy            : return "Transfer Busy";
+        case UPnPResult_CDS_NoSuchFileTransfer      : return "No Such File Transfer";
+        case UPnPResult_CDS_NoSuchDestRes           : return "No Such Destination Resource";
+        case UPnPResult_CDS_DestResAccessDenied     : return "Destination Resource Access Denied";
+        case UPnPResult_CDS_CannotProcessRequest    : return "Cannot Process The Request";
+
+        //case UPnPResult_CMGR_IncompatibleProtocol     = 701,
+        //case UPnPResult_CMGR_IncompatibleDirections   = 702,
+        //case UPnPResult_CMGR_InsufficientNetResources = 703,
+        //case UPnPResult_CMGR_LocalRestrictions        = 704,
+        //case UPnPResult_CMGR_AccessDenied             = 705,
+        //case UPnPResult_CMGR_InvalidConnectionRef     = 706,
+        case UPnPResult_CMGR_NotInNetwork           : return "Not In Network";
+        case UPnPResult_MythTV_NoNamespaceGiven:      return "No Namespace Given";
+        case UPnPResult_MythTV_XmlParseError        : return "XML Parse Error";
+    }
+
+    return "Unknown";
+}
 
 static std::array<const MIMETypes,66> g_MIMETypes
 {{
@@ -451,6 +509,26 @@ qint64 HTTPRequest::SendResponse( void )
     return( nBytes );
 }
 
+void HTTPRequest::SendResponseRedirect(const QString &hostName)
+{
+    m_eResponseType     = ResponseTypeOther;
+    m_nResponseStatus   = 301;
+
+    QStringList sItems = m_sRawRequest.split( ' ' );
+    QString sUrl = "http://" + GetLastHeader( "host" ) + sItems[1];
+    QUrl url( sUrl );
+    QString ipAddress = gCoreContext->GetSettingOnHost
+                            ("BackendServerAddr",hostName,hostName);
+    url.setHost( ipAddress );
+
+    m_mapRespHeaders[ "Location" ] = url.toString();
+
+    LOG(VB_UPNP, LOG_INFO, QString("Sending http redirect to: %1")
+                               .arg(url.toString()));
+
+    SendResponse();
+}
+
 /////////////////////////////////////////////////////////////////////////////
 //
 /////////////////////////////////////////////////////////////////////////////
@@ -696,6 +774,29 @@ void HTTPRequest::FormatErrorResponse( bool  bServerError,
     }
 
     stream.flush();
+}
+
+void HTTPRequest::FormatErrorResponse(UPnPResultCode eCode, const QString &msg)
+{
+    QString sMsg( msg );
+    QString sDetails = "";
+
+    if (m_bSOAPRequest)
+        sDetails = "<UPnPResult xmlns=\"urn:schemas-upnp-org:control-1-0\">";
+
+    if (sMsg.length() == 0)
+        sMsg = UPnPResultCodeDesc(eCode);
+
+    sDetails += QString( "<errorCode>%1</errorCode>"
+                            "<errorDescription>%2</errorDescription>" )
+                    .arg( eCode )
+                    .arg(QString::fromUtf8(QUrl::toPercentEncoding(sMsg)));
+
+    if (m_bSOAPRequest)
+        sDetails += "</UPnPResult>";
+
+    FormatErrorResponse(true, // -=>TODO: Should make this dynamic
+                        "UPnPResult", sDetails);
 }
 
 /////////////////////////////////////////////////////////////////////////////
