+++ mythtv/libs/libmythupnp/upnpcmgr.cpp
@@ -14,7 +14,8 @@
 #include "libmythbase/configuration.h"
 #include "libmythbase/mythlogging.h"
 
-#include "upnp.h"
+#include "httprequest.h"
+#include "upnpresultcode.h"
 
 /////////////////////////////////////////////////////////////////////////////
 //
@@ -155,7 +156,7 @@ bool UPnpCMGR::ProcessRequest( HTTPRequest *pRequest )
                 HandleGetFeatureList( pRequest );
                 break;
             default:
-                UPnp::FormatErrorResponse( pRequest, UPnPResult_InvalidAction );
+                pRequest->FormatErrorResponse(UPnPResult_InvalidAction);
                 break;
         }       
         return true;
@@ -190,8 +191,7 @@ void UPnpCMGR::HandleGetCurrentConnectionInfo( HTTPRequest *pRequest )
 
     if ( nId != 0)
     {
-        UPnp::FormatErrorResponse( pRequest,
-                                   UPnPResult_CMGR_InvalidConnectionRef );
+        pRequest->FormatErrorResponse(UPnPResult_CMGR_InvalidConnectionRef);
         return;
     }
